pyproject.toml
[tool.poetry]
name = "ClauseRadar"
version = "0.1.0"
description = "An AI-powered application to search for contract clauses in PDF documents"
authors = ["Nishat Ahmed <nishat.ahmed@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.24.0"
pinecone = "7.0.1"
pinecone-client = "^2.2.1"
google-generativeai = "^0.8.5"
pdfplumber = "^0.8.1"
numpy = "^1.24.0"
plotly = "^5.14.1"
pyreadline3 = "^3.5.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
